@@ -5,31 +5,32 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.2.3 <4.0.0"
+  sdk: ">=3.8.0 <4.0.0"
 
 dependencies:
+  cached_network_image: ^3.4.1
   cupertino_icons: ^1.0.8
-  dio: ^5.8.0+1
-  easy_localization: ^3.0.7
+  dio: ^5.9.0
+  easy_localization: ^3.0.8
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
   flutter_mobx: ^2.3.0
   flutter_secure_storage: ^9.2.4
-  go_router: ^15.1.2
+  go_router: ^16.2.4
   json_annotation: ^4.9.0
   mobx: ^2.5.0
   provider: ^6.1.5
 
 dev_dependencies:
-  build_runner: ^2.4.15
-  flutter_lints: ^5.0.0
+  build_runner: ^2.7.1
+  flutter_lints: ^6.0.0
   flutter_test:
     sdk: flutter
-  go_router_builder: ^2.9.0
-  json_serializable: ^6.9.0
-  mobx_codegen: ^2.7.0
+  go_router_builder: ^4.1.0
+  json_serializable: ^6.11.1
+  mobx_codegen: ^2.7.4
 
 flutter:
   uses-material-design: true
