@@ -9,8 +9,7 @@ part of 'counter_store.dart';
 // ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic, no_leading_underscores_for_local_identifiers
 
 mixin _$CounterStore on CounterStoreBase, Store {
-  late final _$valueAtom =
-      Atom(name: 'CounterStoreBase.value', context: context);
+  late final _$valueAtom = Atom(name: 'CounterStoreBase.value', context: context);
 
   @override
   int get value {
@@ -25,13 +24,11 @@ mixin _$CounterStore on CounterStoreBase, Store {
     });
   }
 
-  late final _$CounterStoreBaseActionController =
-      ActionController(name: 'CounterStoreBase', context: context);
+  late final _$CounterStoreBaseActionController = ActionController(name: 'CounterStoreBase', context: context);
 
   @override
   void increment() {
-    final _$actionInfo = _$CounterStoreBaseActionController.startAction(
-        name: 'CounterStoreBase.increment');
+    final _$actionInfo = _$CounterStoreBaseActionController.startAction(name: 'CounterStoreBase.increment');
     try {
       return super.increment();
     } finally {
